+++ src/main/java/dev/woori/wooriLearn/domain/auth/controller/AuthController.java
@@ -5,7 +5,6 @@ import dev.woori.wooriLearn.config.response.ApiResponse;
 import dev.woori.wooriLearn.config.response.BaseResponse;
 import dev.woori.wooriLearn.config.response.SuccessCode;
 import dev.woori.wooriLearn.domain.auth.dto.RefreshReqDto;
-import dev.woori.wooriLearn.domain.auth.dto.SignupReqDto;
 import dev.woori.wooriLearn.domain.auth.service.AuthService;
 import dev.woori.wooriLearn.domain.auth.dto.LoginReqDto;
 import lombok.RequiredArgsConstructor;
@@ -25,11 +24,6 @@ import java.security.Principal;
 public class AuthController {
 
     private final AuthService authService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<BaseResponse<?>> signup(@RequestBody SignupReqDto signupReqDto) {
-        return ApiResponse.success(SuccessCode.CREATED, authService.signup(signupReqDto));
-    }
 
     @PostMapping("/login")
     public ResponseEntity<BaseResponse<?>> login(@RequestBody LoginReqDto loginReqDto) {
