+++ src/main/java/dev/woori/wooriLearn/domain/user/controller/UserController.java
@@ -1,7 +1,15 @@
 package dev.woori.wooriLearn.domain.user.controller;
 
+import dev.woori.wooriLearn.config.response.ApiResponse;
+import dev.woori.wooriLearn.config.response.BaseResponse;
+import dev.woori.wooriLearn.config.response.SuccessCode;
+import dev.woori.wooriLearn.domain.user.dto.SignupReqDto;
+import dev.woori.wooriLearn.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -11,4 +19,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class UserController {
 
+    private final UserService userService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<BaseResponse<?>> signup(@RequestBody SignupReqDto signupReqDto) {
+        return ApiResponse.success(SuccessCode.CREATED, userService.signup(signupReqDto));
+    }
 }
