+++ src/main/java/dev/woori/wooriLearn/domain/user/dto/SignupReqDto.java
@@ -1,4 +1,4 @@
-package dev.woori.wooriLearn.domain.auth.dto;
+package dev.woori.wooriLearn.domain.user.dto;
 
 public record SignupReqDto(
     String userId,
